DEADiff/requirements.txt
pytorch_lightning==1.8.6
torchmetrics>=0.7.0
albumentations==0.4.3
pudb==2019.2
imageio==2.9.0
imageio-ffmpeg==0.4.2
omegaconf==2.1.1
test-tube>=0.7.5
streamlit>=0.73.1
einops==0.4.1
torch-fidelity==0.3.0
transformers==4.28.0
accelerate==0.20.3
timm==0.5.4
opencv-python-headless==4.6.0.66
invisible-watermark==0.1.5
matplotlib==3.5.3
tqdm==4.64.1
scipy==1.7.3
ipykernel==6.16.2
PyYAML==6.0
jupyter==1.0.0
notebook==6.5.2
pylint==2.10.2
easydict==1.10
scikit-image==0.19.3
lmdb==1.4.0
ninja==1.10.0
onnxruntime-gpu==1.11.0
submitit==1.4.5
pandas==1.3.5
thop==0.1.1.post2209072238
click==8.1.3
ftfy==6.1.1
kornia==0.6.8
einops-exts==0.0.3
entmax==1.1
k-diffusion==0.0.12
fairscale==0.4.6
open_clip_torch==2.7.0
diffusers==0.16.0
numpy==1.22.0
safetensors==0.4.2
gradio==3.50.2
xformers==0.0.17
httpx==0.23.3
bson
spacy
pycocoevalcap
iopath
decord
webdataset